Standardize Assignments
A quick first pass at standardizing assignments (it works -- although it is not documented in the API).

It looks like canvas-api is going to need some loving, so that it can handle multiple instances of the Canvas API with separate pagination.

This may be the moment when it becomes objectified!

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>